requirements.txt
requests
langchain
langchainhub
langchain-tavily
langgraph
langchain-google-genai
fuzzywuzzy
python-levenshtein
python-dotenv
Flask
certifi